LayoutTests/platform/ios-simulator/fast/events/tab-focus-link-in-canvas-expected.txt

Ensures that you can tab to focus a link inside a canvas subtree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.activeElement.id is "before"
FAIL document.activeElement.id should be canvas_link. Was .
PASS successfullyParsed is true

TEST COMPLETE
Before 
